+++ spec/unit/egis/table_ddl_generator_spec.rb
@@ -71,7 +71,7 @@ RSpec.describe Egis::TableDDLGenerator do
     describe 'table format' do
       subject { strip_whitespaces(generator.create_table_sql(table)) }
 
-      context 'when given tsv format' do
+      context 'when given tsv format preset' do
         let(:format) { :tsv }
 
         let(:expected_query) do
@@ -93,7 +93,7 @@ RSpec.describe Egis::TableDDLGenerator do
         it { is_expected.to eq(expected_query) }
       end
 
-      context 'when given csv format' do
+      context 'when given csv format preset' do
         let(:format) { :csv }
 
         let(:expected_query) do
@@ -115,8 +115,8 @@ RSpec.describe Egis::TableDDLGenerator do
         it { is_expected.to eq(expected_query) }
       end
 
-      context 'when given ORC format' do
-        let(:format) { :orc }
+      context 'when given orc_legacy format preset' do
+        let(:format) { :orc_legacy }
 
         let(:expected_query) do
           strip_whitespaces <<~SQL
@@ -137,8 +137,8 @@ RSpec.describe Egis::TableDDLGenerator do
         it { is_expected.to eq(expected_query) }
       end
 
-      context 'when given table serde' do
-        let(:format) { {serde: 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe'} }
+      context 'when given orc format preset' do
+        let(:format) { :orc }
 
         let(:expected_query) do
           strip_whitespaces <<~SQL
@@ -151,38 +151,12 @@ RSpec.describe Egis::TableDDLGenerator do
               `dth` int,
               `type` string
             )
-            ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe'
-            LOCATION '#{location}';
-          SQL
-        end
-
-        it { is_expected.to eq(expected_query) }
-      end
-
-      context 'when given table serde with serde properties' do
-        let(:format) do
-          {
-            serde: 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe',
-            serde_properties: {'serialization.format' => ',', 'field.delim' => ','}
-          }
-        end
-
-        let(:expected_query) do
-          strip_whitespaces <<~SQL
-            CREATE EXTERNAL TABLE #{table_name} (
-              `id` int,
-              `message` string,
-              `time` timestamp
-            )
-            PARTITIONED BY (
-              `dth` int,
-              `type` string
-            )
-            ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe'
+            ROW FORMAT SERDE 'org.apache.hadoop.hive.ql.io.orc.OrcSerde'
             WITH SERDEPROPERTIES (
-              'serialization.format' = ',',
-              'field.delim' = ','
+              'orc.column.index.access' = 'false'
             )
+            STORED AS INPUTFORMAT 'org.apache.hadoop.hive.ql.io.orc.OrcInputFormat'
+            OUTPUTFORMAT 'org.apache.hadoop.hive.ql.io.orc.OrcOutputFormat'
             LOCATION '#{location}';
           SQL
         end
@@ -190,13 +164,8 @@ RSpec.describe Egis::TableDDLGenerator do
         it { is_expected.to eq(expected_query) }
       end
 
-      context 'when given table serde with serde input format' do
-        let(:format) do
-          {
-            serde: 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe',
-            serde_input_format: 'org.apache.hadoop.mapred.TextInputFormat'
-          }
-        end
+      context 'when format is a string' do
+        let(:format) { 'CUSTOM FORMAT' }
 
         let(:expected_query) do
           strip_whitespaces <<~SQL
@@ -209,8 +178,7 @@ RSpec.describe Egis::TableDDLGenerator do
               `dth` int,
               `type` string
             )
-            ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe'
-            STORED AS INPUTFORMAT 'org.apache.hadoop.mapred.TextInputFormat'
+            ROW FORMAT CUSTOM FORMAT
             LOCATION '#{location}';
           SQL
         end
@@ -218,35 +186,7 @@ RSpec.describe Egis::TableDDLGenerator do
         it { is_expected.to eq(expected_query) }
       end
 
-      context 'when given table serde with serde output format' do
-        let(:format) do
-          {
-            serde: 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe',
-            serde_output_format: 'org.apache.hadoop.hive.ql.io.HiveIgnoreKeyTextOutputFormat'
-          }
-        end
-
-        let(:expected_query) do
-          strip_whitespaces <<~SQL
-            CREATE EXTERNAL TABLE #{table_name} (
-              `id` int,
-              `message` string,
-              `time` timestamp
-            )
-            PARTITIONED BY (
-              `dth` int,
-              `type` string
-            )
-            ROW FORMAT SERDE 'org.apache.hadoop.hive.serde2.lazy.LazySimpleSerDe'
-            OUTPUTFORMAT 'org.apache.hadoop.hive.ql.io.HiveIgnoreKeyTextOutputFormat'
-            LOCATION '#{location}';
-          SQL
-        end
-
-        it { is_expected.to eq(expected_query) }
-      end
-
-      context 'when given an unsupported format' do
+      context 'when given an unsupported format preset' do
         let(:format) { :unknown_format }
 
         it 'raises an error' do